# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Suv yetkazib berish API",
        "description": "Mobil ilovalar uchun suv yetkazib berish tizimining rasmiy API hujjatlari. Barcha endpointlar Uzbek tilida izohlangan bo‘lib, mijoz (user) va xodim (director/courier) ilovalari uchun mo‘ljallangan.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Asosiy API server (Laravel backend ichidagi /api prefiksi bilan)"
        }
    ],
    "paths": {
...
        "/v1/user/auth/send-otp": {
            "post": {
                "tags": [
                    "Mijoz – Autentifikatsiya"
                ],
                "summary": "Mijozga OTP SMS yuborish",
                "description": "Berilgan telefon raqami bo‘yicha mijozni yaratadi (agar mavjud bo‘lmasa) va unga bir martalik tasdiqlash kodi (OTP) yuboradi. Hozircha demo rejimda 123456 kod ishlatiladi.",
                "operationId": "0f6734d8cfb4d6ca14a03f67e3dc37ef",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone"
                                ],
                                "properties": {
                                    "phone": {
                                        "description": "Mijozning telefon raqami, xalqaro formatda (+998...)",
                                        "type": "string",
                                        "example": "+998901234567"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kod yuborildi"
                    },
                    "422": {
                        "description": "Valdatsiya xatosi (telefon kiritilmagan yoki noto‘g‘ri format)"
                    }
                }
            }
        },
        "/v1/user/auth/verify-otp": {
            "post": {
                "tags": [
                    "Mijoz – Autentifikatsiya"
                ],
                "summary": "OTP kodni tasdiqlash va token olish",
                "description": "Mijoz kiritgan SMS kodi to‘g‘ri bo‘lsa, unga Sanctum access token qaytaradi. Shuningdek FCM token va qurilma maʼlumotlari ham saqlanishi mumkin.",
                "operationId": "3cdd8eca54285dd29622209880f0d91e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "phone",
                                    "code"
                                ],
                                "properties": {
                                    "phone": {
                                        "description": "Mijoz telefon raqami",
                                        "type": "string",
                                        "example": "+998901234567"
                                    },
                                    "code": {
                                        "description": "SMS orqali yuborilgan 6 xonali kod",
                                        "type": "string",
                                        "example": "123456"
                                    },
                                    "fcm_token": {
                                        "description": "Push bildirishnomalar uchun Firebase token (ixtiyoriy)",
                                        "type": "string",
                                        "example": "fcm-token-here",
                                        "nullable": true
                                    },
                                    "device_type": {
                                        "description": "Qurilma turi: android yoki ios",
                                        "type": "string",
                                        "example": "android",
                                        "nullable": true
                                    },
                                    "device_name": {
                                        "description": "Qurilmaning nomi (modeli)",
                                        "type": "string",
                                        "example": "Samsung A54",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token va foydalanuvchi maʼlumotlari muvaffaqiyatli qaytarildi"
                    },
                    "422": {
                        "description": "Kod xato yoki muddati tugagan"
                    }
                }
            }
        },
        "/v1/user/profile": {
            "get": {
                "tags": [
                    "Mijoz – Profil"
                ],
                "summary": "Joriy mijoz profilini olish",
                "operationId": "029d09682c281a786413cdfa85ea1fff",
                "responses": {
                    "200": {
                        "description": "Mijoz profili maʼlumotlari"
                    },
                    "401": {
                        "description": "Token noto‘g‘ri yoki yuborilmagan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/user/auth/check": {
            "get": {
                "tags": [
                    "Mijoz – Autentifikatsiya"
                ],
                "summary": "Token holatini tekshirish",
                "description": "Berilgan Bearer token orqali foydalanuvchi hali aktiv ekanligini tekshiradi.",
                "operationId": "5de530485ec7336d71e48be5480e59f6",
                "responses": {
                    "200": {
                        "description": "Token aktiv va foydalanuvchi maʼlumotlari qaytarildi"
                    },
                    "401": {
                        "description": "Token noto‘g‘ri yoki muddati tugagan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/v1/user/profile/update": {
            "post": {
                "tags": [
                    "Mijoz – Profil"
                ],
                "summary": "Mijoz profilini yangilash",
                "operationId": "b242ab76b50c3f14091905ee42ec9bee",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Yangi ism, kamida 3 ta harf",
                                        "type": "string",
                                        "example": "Ali Valiyev"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil muvaffaqiyatli yangilandi"
                    },
                    "422": {
                        "description": "Valdatsiya xatosi"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Bearer token bilan autentifikatsiya. Format: \\\"Bearer {token}\\\"",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}